+++ src/main/java/com/webank/webase/sign/util/CommonUtils.java
@@ -22,6 +22,7 @@ import org.fisco.bcos.sdk.v3.crypto.signature.ECDSASignatureResult;
 import org.fisco.bcos.sdk.v3.crypto.signature.SM2SignatureResult;
 import org.fisco.bcos.sdk.v3.crypto.signature.SignatureResult;
 import org.fisco.bcos.sdk.v3.model.CryptoType;
+import org.fisco.bcos.sdk.v3.utils.Hex;
 import org.fisco.bcos.sdk.v3.utils.Numeric;
 import org.springframework.validation.BindingResult;
 import com.webank.webase.sign.enums.CodeMessageEnums;
@@ -55,7 +56,7 @@ public class CommonUtils {
         } else {
             byteArr = sigResult2ByteArrECDSA((ECDSASignatureResult) signatureResult);
         }
-        return Numeric.toHexString(byteArr, 0, byteArr.length, false);
+        return Hex.toHexString(byteArr);
     }
 
     private static byte[] sigResult2ByteArrGuomi(SM2SignatureResult signatureResult) {