+++ Changelog.md
@@ -1,4 +1,4 @@
-### v2.0.0-rc1(2021-12-09)
+### lab-rc1(2021-12-09)
 
 **Add**
 - 适配兼容FISCO BCOS v3.0.0
@@ -6,7 +6,7 @@
 
 **兼容性**
 - 支持FISCO-BCOS v3.0.0 及以上版本
-- WeBASE-Front v2.0.0-rc1
+- WeBASE-Front lab-rc1
 详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/lab/)。
 
 ### v1.5.3(2021-09-27)