+++ Changelog.md
@@ -1,3 +1,15 @@
+### lab-rc2(2022-03-30)
+
+**Add**
+- 升级springboot 1.x到2.x
+
+
+**兼容性**
+- 支持FISCO-BCOS v3.0.0 及以上版本
+- WeBASE-Front lab-rc1及以上版本
+详细了解,请阅读[**技术文档**](https://webasedoc.readthedocs.io/zh_CN/lab/)。
+
+
 ### lab-rc1(2021-12-09)
 
 **Add**